Simplify dynamic workspace hierarchy.
Common suite-wise helper functions are all defined in `suites/dataset/index.js`. Tab components are now much simpler.
