+++ meteor-app/imports/ui/pages/dynamic-workspace/suites/dataset/tabs/metadata/index.js
@@ -5,17 +5,18 @@ import {
   MarkDownRenderer,
 } from '/imports/ui/helpers';
 
-import TabBaseClass from '../BaseClass';
+import TabComponent from '../../TabComponent';
 
 export default
-class MetadataTab extends TabBaseClass {
+class MetadataTab extends TabComponent {
+  static tabName = 'metadata';
   static tabIcon = null;
   static tabLabel = 'Metadata';
   static requiredProps = [
     'metadata',
   ];
 
-  renderBody () {
+  render () {
     const {
       metadata: metadataField,
     } = this.props;
